Raise a RequestError on error with requests
The catches the error from the requests library
and raises the base error of itunesiap.  I don't
love the error handling message and think it could
be improved.

Improve error reporting on requests error

When raising the error, the message says that
there is an error, make requests error available
in args.